move NinjaWriter to a build lib
